SageMath version 8.8.beta5, Release Date: 2019-05-11

* tag '8.8.beta5': (3695 commits)
  Updated SageMath version to 8.8.beta5
  polymake: update MongoDB information for linux distros
  polymake: Update info on macOS
  polymake: Mention MOngoDB for polydb
  Mention libterm-readline-gnu-perl
  Mention libterm-readkey-perl
  update docs, as discussed on #24905
  leave GMP's spkg-configure.m4 as it was
  Change format of a 'not tested' doc test annotation
  remove an obsolete note in schemes, pep cleanup of zariski_vankampen
  spkg-config for pcre
  get rid of dollars in elliptic curves
  spkg-configure for perl_term_readline_gnu
  trac 27794: py3 doctests in sets
  Trac #27779: Test the matrix result only for ...
  Fix error message in doctest for new pynormaliz version
  Trac #27779: Add `# abs tol 1e-11` tag to doctest
  Trac #27779: Add missing `# long time` flag
  another harvest of typos
  trac 2693 resultants for polynomials over inexact rings
  ...

+++ src/sage/schemes/affine/affine_point.py
@@ -479,7 +479,7 @@ class SchemeMorphism_point_affine_finite_field(SchemeMorphism_point_affine_field
         """
         p = self.codomain().base_ring().order()
         N = self.codomain().ambient_space().dimension_relative()
-        return sum(hash(self[i])*p**i for i in range(N))
+        return int(sum(hash(self[i])*p**i for i in range(N)))
 
     def orbit_structure(self, f):
         r"""
